requirements.txt
playwright
requests
moviepy
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
python-dotenv
pytest
